requirements.txt
flask==2.2.3
flask-cors==3.0.10
stripe==5.4.0
python-dotenv==1.0.0